pyproject.toml
[project]
name = "adp"
version = "0.1.0"
description = "Agentic Decomposition Pipeline — large model decomposes, small model executes"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "rich>=13.7.0",
    "prompt_toolkit>=3.0.43",
    "pydantic>=2.6.0",
    "networkx>=3.3",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14",
]

[project.scripts]
adp = "adp.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
